==> evaluation/6.cluster_relatio_evaluation.ipynb <==
import pandas as pd
import json
import pickle
import joblib

import matplotlib.pyplot as plt
import seaborn as sns

import sys 
sys.path.append('../')
from fast_tsne.fast_tsne import fast_tsne

import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

from eval_canarex import calculate_mse

## Cluster results

### Textrank

path='relatio_test/clust_100/'

cpath = '{}/kmeans_all.pkl'.format(path)
narratives_path='{}/final_narratives.jsonl'.format(path)
result_path = '{}/results.jsonl'.format(path)
eval_path = '{}/eval_output.csv'.format(path)

ctype='true_text'

### Load scores

df = pd.read_pickle(cpath)

df.head(2)

df = df[df.score == df.score.max()]
df = df.explode(['dist', 'labels'])
df = df.reset_index(drop = True)

df.shape

df.labels.nunique()

df.head(2)

df.shape

### Merge narratives

narratives_path

chunks = pd.read_json(narratives_path, 
                      lines=True, chunksize=10000)
narratives = None
for chunk in chunks:
    if narratives is None:
        narratives = chunk
    else:
        narratives = pd.concat([narratives, chunk], axis=0)

narratives.shape

narratives.index

final = pd.concat([narratives, df.reset_index(drop = True)],
                  axis = 1)

final = final.rename(columns = {'sentence': 'text'})

final.shape

df = pd.read_json('{}/synthetic_test_data_{}.jsonl'.format('synthetic_data', ctype), 
                  lines = True, orient='records')

orig_text = df[['id', 'sentence_id', 'text', 'noise', 'synthetic_label', 'time']].drop_duplicates()


final.shape

final = pd.merge(final, orig_text)

final.shape

# 1 = triangle (18: kevin rudd)
# 2 = square (24: climate change)
# 3 = random wave (63: Indeginous people)

final[final.noise == False].groupby(['synthetic_label', 'labels']).\
size().\
reset_index(name='l').\
sort_values(['synthetic_label', 'l'], ascending=[True, False]).\
groupby(['synthetic_label']).\
head(3)

### Calculate MSE

ctype

df = pd.read_json('{}/synthetic_test_data_{}.jsonl'.format('synthetic_data', ctype), 
                  lines = True, orient='records')

final = pd.read_json(result_path, lines = True, orient='records')

final.labels.unique()

eval_df = calculate_mse(df, final)

eval_path

eval_df.to_csv(eval_path, index = False)

eval_df

eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]

### Save output

eval_path

eval_df.to_csv(eval_path, index = False)

### Create label mapping

label_mapping = eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]
label_mapping = label_mapping[['orig_clust', 'pred_clust']]
label_mapping = label_mapping.rename(columns = {'pred_clust': 'labels',
                                      'orig_clust':'new_labels'})

label_mapping

# column, new_labels: mapped original cluster assignments (3 clusters)
# column, labels: new cluster assignments (N clusters)

# Map the rest of the (N-3) labels to a single cluster (as it is not considered recovery)
# Used for plotting graphs
pending_labels = deque()
pending_labels.extend(set(range(0, final.labels.nunique())) - set(label_mapping.labels.tolist()))

label_mapping_dict = []
for i in range(4, 4 + len(pending_labels)):
    label_mapping_dict.append({'new_labels': 4, 'labels':pending_labels.pop()})

# Add pending labels
label_mapping = pd.concat([label_mapping, pd.DataFrame(label_mapping_dict)], axis = 0).reset_index(drop = True)

# narrative cluster assignments mapped to orignial synthetic cluster
# Cluster 1 = triangle (kevin rudd)
# Cluster 2 = square (climate change)
# Cluster 3 = random wave (Indeginous people)
# Cluster 4 = Other new cluster identified (we treat this as noise)
label_mapping

final = pd.merge(final, label_mapping)

### Plot narrative clusters

cluster_labels = {1:'Cluster 1: Kevin Rudd', #18, #synthetic label 1, Kevin Rudd
2:'Cluster 2: Climate change', # 24 #synthetic label 2 Climate change
3:'Cluster 3: Indeginous people', #63 #synthetic label 3, Indeginous people
4: 'Cluster 4: New cluster'}

colour = {'1':'#abd9e9',
'2':'#fdae61',
'3': '#D2AB3C',
'4': '#5ab4ac'}

test = final[final.noise == False]\
[['sentence_id','time', 'text', 'synthetic_label', 'noise', 'new_labels']].drop_duplicates()
test = test.groupby(['time', 'new_labels']).size().reset_index(name='l')
test = test.sort_values(by=['new_labels'], key=lambda x: x.map(cluster_labels))

fig, ax = plt.subplots(figsize=(8,6))
for label in test.new_labels.unique():
    ax.scatter(test[test.new_labels == label].time.tolist(), 
               test[test.new_labels == label].l.tolist(), 
               c=[colour[str(i)] for i in test[test.new_labels == label].new_labels.tolist()], 
               s=15, label=cluster_labels[label],
               alpha=1, edgecolors='none')

ax.legend(title='Topics', loc=1, prop={'size': 12}, 
          title_fontsize=15, markerscale=2, framealpha=0.5)
ax.grid(False)
ax.set_ylabel("No of sentences", size=14)
ax.set_xlabel("Time", size=14)
ax.grid(False)
plt.show()

### Compare clusters

import pandas as pd

# All clusters
path='relatio_test/clust_40'
eval_df = pd.read_csv('{}/eval_output.csv'.format(path))

eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]

path='relatio_test/clust_100'
eval_df = pd.read_csv('{}/eval_output.csv'.format(path))

eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]
