==> evaluation/7.cluster_topN_evaluation.ipynb <==
import pandas as pd
import json
import pickle
import joblib

import matplotlib.pyplot as plt
import seaborn as sns

import sys 
sys.path.append('../')
from fast_tsne.fast_tsne import fast_tsne

import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

path='canarex_coref_text/topN'
# path='canarex_coref_text_no_split_sentences/topN'
ctype='coref_text'

# path='canarex_true_text/topN'
# path='canarex_true_text_no_split_sentences/topN'
# path='relatio_test/clust_40/topN'
# ctype='true_text'

cpath = '{}/narratives_kmeans2_topN.pkl'.format(path)
narratives_path='{}/final_narratives.jsonl'.format(path)
result_path = '{}/results.jsonl'.format(path)
eval_path = '{}/eval_output.csv'.format(path)



### Load scores

topN = pd.read_pickle(cpath)

topN[topN.score == topN.score.max()]

# get the top N sentences and K based on score
N = topN[topN.score == topN.score.max()].iloc[0].topN
k = topN[topN.score == topN.score.max()].iloc[0].k

N

topN = topN[topN.score == topN.score.max()]
topN = topN.explode(['dist', 'labels']).reset_index(drop = True)

### Load document level clustering

k2 = pd.read_pickle('{}/kmeans2.pkl'.format(path))
k2.drop(columns='text').head(2)

topn = k2.groupby(['id', 'labels_2']).size().reset_index(name='cl')
topn.loc[:, 'maxl'] = topn.groupby('id')['cl'].transform(max)
topn = topn[topn.cl == topn.maxl].drop(columns='maxl')
topn = pd.merge(k2, topn.drop(columns='cl'))

# get the number of narratives b/w 1-5 based on max value
if N == 1:
    print('N = 1')
    topn = topn.sort_values(['id', 'labels_2', 'dist_2'], ascending=True).groupby('id').head(1)
else:
    topn = topn.sort_values(['id', 'labels_2', 'dist_2'], ascending=True).groupby(['id', 'narrative']).head(1)
    topn = topn.sort_values(['id', 'labels_2', 'dist_2'], ascending=True).groupby('id').head(N)

topn = topn.reset_index(drop = True)

topn.shape

topN.shape

topn = pd.concat([topn, topN], axis = 1)

topn.shape

topn.index

topn.columns

topn.to_json(result_path, lines = True, orient='records')

### Calculate MSE

from eval_canarex import calculate_mse

df = pd.read_json('{}/synthetic_test_data_{}.jsonl'.format('synthetic_data', ctype), 
                  lines = True, orient='records')

final = pd.read_json(result_path, lines = True, orient='records')

eval_path

eval_df = calculate_mse(df, final)

eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]

eval_df.to_csv(eval_path, index = False)
