==> evaluation/9.ab-testing.ipynb <==
import numpy as np
from statsmodels.stats import power as pwr
from scipy.stats import chi2_contingency, chi2 
from sklearn import metrics

import pandas as pd
import json
import pickle
import joblib

from collections import deque

def map_label(t):
    # Gives different label mapping w.r.t count and MSE. 
    # We use count here as we are concerned with max recovery
    label_mapping = t[t.noise == False].groupby(['synthetic_label', 'labels']).\
    size().\
    reset_index(name='l').\
    sort_values(['synthetic_label', 'l'], ascending=[True, False]).\
    groupby(['synthetic_label']).\
    head(1)
    label_mapping = label_mapping[['synthetic_label', 'labels']].rename(columns = {'synthetic_label': 'new_labels'})


    # Map the rest of the (N-3) labels to a single cluster (as it is not considered recovery)
    # Used for plotting graphs
    pending_labels = deque()
    pending_labels.extend(set(range(0, t.labels.nunique())) - set(label_mapping.labels.tolist()))

    label_mapping_dict = []
    for i in range(4, 4 + len(pending_labels)):
        label_mapping_dict.append({'new_labels': 4, 'labels':pending_labels.pop()})

    # Add pending labels
    label_mapping = pd.concat([label_mapping, pd.DataFrame(label_mapping_dict)], axis = 0).reset_index(drop = True)
    t = pd.merge(t, label_mapping)
    
    return t



def map_label_by_mse(eval_df, t):
    label_mapping = eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]
    label_mapping = label_mapping[['orig_clust', 'pred_clust']].\
        rename(columns={'orig_clust':'new_labels',
                        'pred_clust':'labels'})
    
    # Map the rest of the (N-3) labels to a single cluster (as it is not considered recovery)
    # Used for plotting graphs
    pending_labels = deque()
    pending_labels.extend(set(range(0, t.labels.nunique())) - set(label_mapping.labels.tolist()))

    label_mapping_dict = []
    for i in range(4, 4 + len(pending_labels)):
        label_mapping_dict.append({'new_labels': 4, 'labels':pending_labels.pop()})

    # Add pending labels
    label_mapping = pd.concat([label_mapping, pd.DataFrame(label_mapping_dict)], axis = 0).reset_index(drop = True)
    t = pd.merge(t, label_mapping)
    
    return t


def map_clusters(t):
    # calculate cluster sizes
    test_clust_count = t[['time','id','synthetic_label', 
                   'new_labels', 'noise', 'text', 'sentence_id']].drop_duplicates().\
    groupby(['synthetic_label', 'new_labels', 'noise']).size().reset_index(name='l')

    # get cluster size for noise (cluster 4)
    noise_clust = test_clust_count[(test_clust_count.noise == True)].\
        groupby('new_labels').l.sum().reset_index()
    noise_clust['synthetic_label'] = 4
    
    # get cluster size without noise (cluster 1-3)
    test_clust_count = test_clust_count[test_clust_count.noise == False].drop(columns='noise')
    # merge
    test_clust_count = pd.concat([test_clust_count, noise_clust]).reset_index(drop = True)

    # label cluster
    test_clust_count['synthetic_label'] = 'C' + test_clust_count['synthetic_label'].astype(str)
    test_clust_count['new_labels'] = 'C' + test_clust_count['new_labels'].astype(str)
    
    # repeat row assignment mapping according to the size of cluster
    test_clust_count = test_clust_count.loc[test_clust_count.index.repeat(test_clust_count.l)]
    test_clust_count = test_clust_count.drop(columns='l')
    
    return(test_clust_count)



def get_proportions(synthetic_data, recovered_data):
    # size of original cluster
    op = synthetic_data[['time','id', 'synthetic_label', 'noise', 'text', 'sentence_id']].drop_duplicates()
    op = op[(op.noise == False)].\
        groupby(['synthetic_label', 'time']).size().reset_index(name='l')

    # size of recovered cluster
    test_count = recovered_data[['time','id','synthetic_label', 
                       'new_labels', 'noise', 'text', 'sentence_id']].drop_duplicates()
    tp = test_count[(test_count.noise == False) & 
             (test_count.synthetic_label == test_count.new_labels)].\
        groupby(['time', 'new_labels']).size().reset_index(name='l')
    
    tp = pd.merge(tp, op, 
             left_on = ['time', 'new_labels'],
             right_on = ['time', 'synthetic_label'])
    tp['prop'] = tp.l_x/tp.l_y
    
    result_prop = tp.groupby(['new_labels'])['prop'].mean()
    
    return(result_prop)

### Precision/recall for CANarEx

path='canarex_coref_text'
result_path = '{}/results.jsonl'.format(path)
eval_path = '{}/eval_output.csv'.format(path)

ctype='coref_text'

df = pd.read_json('{}/synthetic_test_data_{}.jsonl'.format('synthetic_data', ctype), 
                  lines = True, orient='records')

final = pd.read_json(result_path, lines = True)

eval_df = pd.read_csv(eval_path)
eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]

label_mapping = final[final.noise == False].groupby(['synthetic_label', 'labels']).\
size().\
reset_index(name='l').\
sort_values(['synthetic_label', 'l'], ascending=[True, False]).\
groupby(['synthetic_label']).\
head(1)

label_mapping

final = map_label(final)

test_clust_count = map_clusters(final)

test_clust_count.head(5)

print(metrics.confusion_matrix(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

print(metrics.classification_report(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

#### Proportion recovered

canarex_prop = get_proportions(df, final)

### Precision/recall for Relatio

path='relatio_test/clust_100/'

result_path = '{}/results.jsonl'.format(path)
eval_path = '{}/eval_output.csv'.format(path)

ctype='true_text'

df = pd.read_json('{}/synthetic_test_data_{}.jsonl'.format('synthetic_data', ctype), 
                  lines = True, orient='records')

final = pd.read_json(result_path, lines = True)

final = map_label(final)

test_clust_count = map_clusters(final)

print(metrics.confusion_matrix(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

print(metrics.classification_report(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

#### Proportion recovered

relatio_prop = get_proportions(df, final)

#### MSE labels

# note: MSE labels differ from count labels
eval_df = pd.read_csv(eval_path)
eval_df[eval_df.mse == eval_df.groupby(['orig_clust'])['mse'].transform('min')]

final = pd.read_json(result_path, lines = True)

final = map_label_by_mse(eval_df, final)

test_clust_count = map_clusters(final)

print(metrics.confusion_matrix(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

print(metrics.classification_report(test_clust_count.synthetic_label.tolist(), 
                               test_clust_count.new_labels.tolist()))

### A/B testing

canarex_prop

relatio_prop

from scipy.stats import ttest_ind

ttest_ind(canarex_prop, 
      relatio_prop)
